node: move unit tests from old-style nodes to new-style ones
